securize patch and delete user

+++ src/users/users.controller.ts
@@ -1,5 +1,6 @@
-import { Controller, Post, Get, Patch, Body, Param, ParseIntPipe, Delete } from '@nestjs/common';
+import { Controller, Get, Patch, Delete, Param, Body, ParseIntPipe, UseGuards, Request, ForbiddenException } from '@nestjs/common';
 import { UsersService } from './users.service';
+import { AuthGuard } from '@nestjs/passport';
 import { User } from 'src/users/entities/user.entity';
 
 @Controller('users')
@@ -18,18 +19,27 @@ export class UsersController {
     return this.usersService.findOne(id);
   }
 
-  // PATCH /users/:id
+  // PATCH /users/:id (protégé)
+  @UseGuards(AuthGuard('jwt'))
   @Patch(':id')
   update(
     @Param('id', ParseIntPipe) id: number,
     @Body() updateData: Partial<User>,
+    @Request() req,
   ) {
+    if (req.user.userId !== id) {
+      throw new ForbiddenException('You can only update your own profile');
+    }
     return this.usersService.update(id, updateData);
   }
 
-  // DELETE /users/:id
+  // DELETE /users/:id (protégé)
+  @UseGuards(AuthGuard('jwt'))
   @Delete(':id')
-  remove(@Param('id', ParseIntPipe) id: number) {
+  remove(@Param('id', ParseIntPipe) id: number, @Request() req) {
+    if (req.user.userId !== id) {
+      throw new ForbiddenException('You can only delete your own profile');
+    }
     return this.usersService.remove(id);
   }
 }
